% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{canonical}
\alias{canonical}
\title{DNA complementary function}
\usage{
canonical(seq, minimal = TRUE)
}
\arguments{
\item{seq}{The DNA or RNA sequence}

\item{minimal}{Return the lexicographical minimal sequence? Use logical. Default: TRUE}
}
\value{
The complementary sequence
}
\description{
This function returns the complementary sequence for DNA or RNA
}
