% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementary.R
\name{complementary}
\alias{complementary}
\title{DNA complementary function}
\usage{
complementary(seq, dna = TRUE)
}
\arguments{
\item{seq}{The DNA or RNA sequence}

\item{dna}{Is it a DNA sequence, or RNA? Use logical. Default: TRUE}
}
\value{
The complementary sequence
}
\description{
This function returns the complementary sequence for DNA or RNA
}
